+++ simplicio_loop/__init__.py
@@ -13,6 +13,6 @@ try:
     try:
         __version__ = _v("simplicio-loop")
     except PackageNotFoundError:
-        __version__ = "3.38.34"
+        __version__ = "3.38.35"
 except Exception:  # pragma: no cover
-    __version__ = "3.38.34"
+    __version__ = "3.38.35"